Do not drop the result of protocol detection
In MousePickProtocol() with protocol PROT_AUTO we probe for the protocol to
use but drop the result in most cases.  This was causing DEVICE_INIT and
DEVICE_ON to fail to be called with the VUID protocol.  Git history suggests
that this code was originally meant to cover both PS/2 auto-detection and OS-
specific detection, but that only the first case was implemented at the time.
Now that only the second is needed dropping the result to keep the protocol
as PROT_AUTO is presumably no longer useful and seems to actively breaking
things.

Signed-off-by: Michael Thayer <michael.thayer@oracle.com>
Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>