drop Linux support
Since the standard API on Linux is evdev (and we've got evdev and
libinput drivers for that), there's probably no need for using this
driver on Linux anymore. So, we can drop Linux support here.

Making the build fail on Linux, similar to xf86-input-keyboard.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>
Part-of: <https://gitlab.freedesktop.org/xorg/driver/xf86-input-mouse/-/merge_requests/19>